I init useSetupActions hook.
